+++ include/winstl/filesystem/handles/memory_mapped_file_view_handle.hpp
@@ -1,13 +1,13 @@
 /* /////////////////////////////////////////////////////////////////////////
- * File:        winstl/filesystem/handles/memory_mapped_file_view_handle.hpp
+ * File:    winstl/filesystem/handles/memory_mapped_file_view_handle.hpp
  *
- * Purpose:     Handle::Ref+Wrapper handle-adapter class template for
- *              Windows memory mapped file view handles.
+ * Purpose: Handle::Ref+Wrapper handle-adapter class template for Windows
+ *          memory mapped file view handles.
  *
- * Created:     30th August 2010
- * Updated:     22nd January 2024
+ * Created: 30th August 2010
+ * Updated: 22nd February 2024
  *
- * Home:        http://stlsoft.org/
+ * Home:    http://stlsoft.org/
  *
  * Copyright (c) 2019-2024, Matthew Wilson and Synesis Information Systems
  * Copyright (c) 2010-2019, Matthew Wilson and Synesis Software
@@ -54,9 +54,10 @@
 #ifndef STLSOFT_DOCUMENTATION_SKIP_SECTION
 # define WINSTL_VER_WINSTL_FILESYSTEM_HANDLES_HPP_MEMORY_MAPPED_FILE_VIEW_HANDLE_MAJOR      1
 # define WINSTL_VER_WINSTL_FILESYSTEM_HANDLES_HPP_MEMORY_MAPPED_FILE_VIEW_HANDLE_MINOR      0
-# define WINSTL_VER_WINSTL_FILESYSTEM_HANDLES_HPP_MEMORY_MAPPED_FILE_VIEW_HANDLE_REVISION   6
-# define WINSTL_VER_WINSTL_FILESYSTEM_HANDLES_HPP_MEMORY_MAPPED_FILE_VIEW_HANDLE_EDIT       13
+# define WINSTL_VER_WINSTL_FILESYSTEM_HANDLES_HPP_MEMORY_MAPPED_FILE_VIEW_HANDLE_REVISION   7
+# define WINSTL_VER_WINSTL_FILESYSTEM_HANDLES_HPP_MEMORY_MAPPED_FILE_VIEW_HANDLE_EDIT       14
 #endif /* !STLSOFT_DOCUMENTATION_SKIP_SECTION */
+
 
 /* /////////////////////////////////////////////////////////////////////////
  * includes
@@ -83,6 +84,7 @@
 # include <stlsoft/quality/cover.h>
 #endif /* !STLSOFT_INCL_STLSOFT_QUALITY_H_COVER */
 
+
 /* /////////////////////////////////////////////////////////////////////////
  * namespace
  */
@@ -104,6 +106,7 @@ namespace winstl_project
 # endif /* STLSOFT_NO_NAMESPACE */
 #endif /* !WINSTL_NO_NAMESPACE */
 
+
 /* /////////////////////////////////////////////////////////////////////////
  * classes
  */
@@ -123,6 +126,9 @@ public:
         : memory(p)
         , size(n)
     {}
+#if __cplusplus >= 201103L
+    memory_mapped_file_view_state_t(memory_mapped_file_view_state_t const&) = default;
+#endif
 private:
     void operator =(memory_mapped_file_view_state_t const&);    // copy-assignment proscribed
 };
@@ -216,6 +222,7 @@ private:
     const bool      isOwner;
 };
 
+
 /* /////////////////////////////////////////////////////////////////////////
  * shims
  */
@@ -246,6 +253,7 @@ get_memory_mapped_file_view_handle(
     return (NULL != h) ? h->handle : ss_typename_type_k memory_mapped_file_view_handle<R>::handle_type(NULL, 0u);
 }
 
+
 /* ////////////////////////////////////////////////////////////////////// */
 
 #ifndef WINSTL_NO_NAMESPACE
@@ -257,6 +265,7 @@ get_memory_mapped_file_view_handle(
 } /* namespace stlsoft */
 # endif /* STLSOFT_NO_NAMESPACE */
 #endif /* !WINSTL_NO_NAMESPACE */
+
 
 /* /////////////////////////////////////////////////////////////////////////
  * inclusion control